backend/apps/authentication/uploads/tests/test_upload_user_avatars.py
"""
Unit tests for Avatar Upload functionality - Standalone version
Tests the core logic without Django test framework dependencies
"""

import unittest
import os
import tempfile
from io import BytesIO
from unittest.mock import Mock, patch, MagicMock

# Mock Django modules for standalone testing
import sys
from unittest.mock import MagicMock

# Create mock modules
mock_django = MagicMock()
mock_rest_framework = MagicMock()
mock_django_auth = MagicMock()

sys.modules['django'] = mock_django
sys.modules['django.conf'] = mock_django.conf
sys.modules['django.contrib'] = mock_django.contrib
sys.modules['django.contrib.auth'] = mock_django_auth
sys.modules['django.contrib.auth.models'] = mock_django_auth.models
sys.modules['django.core'] = mock_django.core
sys.modules['django.core.files'] = mock_django.core.files
sys.modules['django.core.files.storage'] = mock_django.core.files.storage
sys.modules['django.core.files.base'] = mock_django.core.files.base
sys.modules['rest_framework'] = mock_rest_framework
sys.modules['rest_framework.views'] = mock_rest_framework.views
sys.modules['rest_framework.parsers'] = mock_rest_framework.parsers
sys.modules['rest_framework.response'] = mock_rest_framework.response
sys.modules['rest_framework.status'] = mock_rest_framework.status
sys.modules['rest_framework.permissions'] = mock_rest_framework.permissions

# Mock status codes
mock_rest_framework.status.HTTP_200_OK = 200
mock_rest_framework.status.HTTP_400_BAD_REQUEST = 400
mock_rest_framework.status.HTTP_401_UNAUTHORIZED = 401
mock_rest_framework.status.HTTP_500_INTERNAL_SERVER_ERROR = 500

try:
    from PIL import Image
    PIL_AVAILABLE = True
except ImportError:
    PIL_AVAILABLE = False
    print("Warning: PIL not available, skipping image creation tests")


class AvatarUploadLogicTest(unittest.TestCase):
    """Test the core logic of avatar upload functionality"""
    
    def test_file_extension_validation(self):
        """Test file extension validation logic"""
        allowed_extensions = ['.jpg', '.jpeg', '.png']
        
        # Test valid extensions
        valid_files = ['image.jpg', 'photo.jpeg', 'avatar.png', 'TEST.JPG', 'file.PNG']
        for filename in valid_files:
            extension = os.path.splitext(filename)[1].lower()
            self.assertIn(extension, allowed_extensions, f"Extension {extension} should be allowed")
        
        # Test invalid extensions
        invalid_files = ['document.gif', 'file.txt', 'image.bmp', 'video.mp4']
        for filename in invalid_files:
            extension = os.path.splitext(filename)[1].lower()
            self.assertNotIn(extension, allowed_extensions, f"Extension {extension} should not be allowed")
    
    def test_file_size_validation(self):
        """Test file size validation logic"""
        max_size = 2 * 1024 * 1024  # 2MB
        
        # Test valid sizes
        valid_sizes = [1024, 50000, 1000000, max_size - 1, max_size]
        for size in valid_sizes:
            self.assertLessEqual(size, max_size, f"Size {size} should be valid")
        
        # Test invalid sizes
        invalid_sizes = [max_size + 1, max_size * 2, max_size * 10]
        for size in invalid_sizes:
            self.assertGreater(size, max_size, f"Size {size} should be invalid")
    
    def test_filename_generation(self):
        """Test unique filename generation logic"""
        import uuid
        
        # Mock uuid generation
        with patch('uuid.uuid4', return_value='test-uuid-12345'):
            filename = f"{uuid.uuid4()}.jpg"
            expected = "test-uuid-12345.jpg"
            self.assertEqual(filename, expected)
    
    def test_avatar_path_construction(self):
        """Test avatar path construction"""
        filename = "test-avatar.jpg"
        avatar_path = f"avatars/{filename}"
        expected_path = "avatars/test-avatar.jpg"
        self.assertEqual(avatar_path, expected_path)
    
    def test_media_url_construction(self):
        """Test media URL construction"""
        media_url = "/media/"
        saved_path = "avatars/test-avatar.jpg"
        avatar_url = f"{media_url}{saved_path}"
        expected_url = "/media/avatars/test-avatar.jpg"
        self.assertEqual(avatar_url, expected_url)
    
    @unittest.skipUnless(PIL_AVAILABLE, "PIL not available")
    def test_image_file_creation(self):
        """Test creating valid image files for testing"""
        # Create a test image
        image = Image.new('RGB', (100, 100), color='red')
        image_io = BytesIO()
        image.save(image_io, format='JPEG')
        image_data = image_io.getvalue()
        
        # Verify image data is not empty
        self.assertGreater(len(image_data), 0)
        
        # Verify image data starts with JPEG header
        self.assertTrue(image_data.startswith(b'\xff\xd8\xff'))
    
    def test_error_message_constants(self):
        """Test error message consistency"""
        error_messages = {
            'no_file': "No file provided.",
            'size_exceeded': "File size exceeds 2MB limit.",
            'invalid_format': "Only JPG and PNG files are allowed.",
            'upload_success': "Avatar uploaded successfully"
        }
        
        # Verify all error messages are non-empty strings
        for key, message in error_messages.items():
            self.assertIsInstance(message, str)
            self.assertGreater(len(message), 0)
            self.assertTrue(message.endswith('.') or key == 'upload_success')
    
    def test_case_insensitive_extension_check(self):
        """Test case-insensitive file extension validation"""
        test_cases = [
            ('image.jpg', True),
            ('image.JPG', True),
            ('image.Jpg', True),
            ('image.jpeg', True),
            ('image.JPEG', True),
            ('image.png', True),
            ('image.PNG', True),
            ('image.gif', False),
            ('image.GIF', False),
            ('document.txt', False),
        ]
        
        allowed_extensions = ['.jpg', '.jpeg', '.png']
        
        for filename, should_be_valid in test_cases:
            extension = os.path.splitext(filename)[1].lower()
            is_valid = extension in allowed_extensions
            self.assertEqual(is_valid, should_be_valid, 
                           f"File {filename} validation result should be {should_be_valid}")


class AvatarUploadMockTest(unittest.TestCase):
    """Test avatar upload with mocked dependencies"""
    
    def setUp(self):
        """Set up mock objects"""
        self.mock_request = Mock()
        self.mock_request.FILES = {}
        self.mock_request.user = Mock()
        self.mock_request.user.username = 'testuser'
        
        # Mock user profile
        self.mock_userprofile = Mock()
        self.mock_userprofile.avatar = None
        self.mock_request.user.userprofile = self.mock_userprofile
    
    def test_no_file_in_request(self):
        """Test handling when no file is provided"""
        # Simulate no file in request
        self.mock_request.FILES = Mock()
        self.mock_request.FILES.get.return_value = None
        
        # The view should detect missing file
        file = self.mock_request.FILES.get('avatar')
        self.assertIsNone(file)
    
    def test_file_size_check(self):
        """Test file size validation"""
        # Mock a large file
        mock_file = Mock()
        mock_file.size = 3 * 1024 * 1024  # 3MB (over limit)
        
        max_size = 2 * 1024 * 1024  # 2MB limit
        is_too_large = mock_file.size > max_size
        
        self.assertTrue(is_too_large)
    
    def test_file_extension_check(self):
        """Test file extension validation"""
        # Mock file with invalid extension
        mock_file = Mock()
        mock_file.name = 'test.gif'
        
        allowed_extensions = ['.jpg', '.jpeg', '.png']
        file_extension = os.path.splitext(mock_file.name)[1].lower()
        is_valid_format = file_extension in allowed_extensions
        
        self.assertFalse(is_valid_format)
    
    def test_successful_response_format(self):
        """Test the format of successful response"""
        # Expected response format
        expected_response = {
            "message": "Avatar uploaded successfully",
            "avatar_url": "/media/avatars/test-avatar.jpg"
        }
        
        # Verify response structure
        self.assertIn('message', expected_response)
        self.assertIn('avatar_url', expected_response)
        self.assertEqual(expected_response['message'], 'Avatar uploaded successfully')
        self.assertTrue(expected_response['avatar_url'].startswith('/media/'))
    
    def test_error_response_format(self):
        """Test the format of error responses"""
        error_responses = [
            {"error": "No file provided."},
            {"error": "File size exceeds 2MB limit."},
            {"error": "Only JPG and PNG files are allowed."},
            {"error": "Upload failed: Storage error"}
        ]
        
        for response in error_responses:
            self.assertIn('error', response)
            self.assertIsInstance(response['error'], str)
            self.assertGreater(len(response['error']), 0)


class FileUtilityTest(unittest.TestCase):
    """Test file utility functions"""
    
    def test_temporary_directory_creation(self):
        """Test creating temporary directories for testing"""
        with tempfile.TemporaryDirectory() as temp_dir:
            self.assertTrue(os.path.exists(temp_dir))
            self.assertTrue(os.path.isdir(temp_dir))
    
    def test_file_path_operations(self):
        """Test file path operations"""
        # Test path joining
        base_path = "/media"
        filename = "avatars/test.jpg"
        full_path = os.path.join(base_path, filename)
        
        # Verify path construction
        self.assertIn("media", full_path)
        self.assertIn("test.jpg", full_path)
    
    def test_extension_extraction(self):
        """Test file extension extraction"""
        test_files = [
            ("image.jpg", ".jpg"),
            ("photo.jpeg", ".jpeg"),
            ("avatar.png", ".png"),
            ("document.pdf", ".pdf"),
            ("file", ""),  # No extension
        ]
        
        for filename, expected_ext in test_files:
            actual_ext = os.path.splitext(filename)[1]
            self.assertEqual(actual_ext, expected_ext)


if __name__ == '__main__':
    # Configure test runner
    unittest.TestLoader.testMethodPrefix = 'test'
    
    # Create test suite
    test_suite = unittest.TestSuite()
    
    # Add test classes
    test_classes = [
        AvatarUploadLogicTest,
        AvatarUploadMockTest,
        FileUtilityTest
    ]
    
    for test_class in test_classes:
        tests = unittest.TestLoader().loadTestsFromTestCase(test_class)
        test_suite.addTests(tests)
    
    # Run tests
    runner = unittest.TextTestRunner(verbosity=2)
    result = runner.run(test_suite)
    
    # Print summary
    print(f"\n{'='*50}")
    print(f"AVATAR UPLOAD TESTS SUMMARY")
    print(f"{'='*50}")
    print(f"Tests run: {result.testsRun}")
    print(f"Failures: {len(result.failures)}")
    print(f"Errors: {len(result.errors)}")
    print(f"Success rate: {((result.testsRun - len(result.failures) - len(result.errors)) / result.testsRun * 100):.1f}%")
    
    if result.failures:
        print(f"\nFAILURES:")
        for test, traceback in result.failures:
            print(f"- {test}: {traceback}")
    
    if result.errors:
        print(f"\nERRORS:")
        for test, traceback in result.errors:
            print(f"- {test}: {traceback}")
    
    # Exit with appropriate code
    exit_code = 0 if result.wasSuccessful() else 1
    print(f"\nTest {'PASSED' if exit_code == 0 else 'FAILED'}")
    exit(exit_code)
