backend/requirements.txt
django
PyMySQL
requests
django-cors-headers
django-crispy-forms
django-filter
django-tables2
django-multiselectfield
django-bootstrap3
django-bootstrap4
django-debug-toolbar
django-redis
django-storages
dj-database-url
djangorestframework
djangorestframework-simplejwt
pytz
whitenoise
numpy 
pandas
pillow
matplotlib
beautifulsoup4
PyYAML
python-dateutil
python-decouple
gunicorn
wordcloud
drf-spectacular
cryptography
python-dotenv
redis
celery
mysqlclient
channels
channels-redis
daphne
weasyprint
markdown
psycopg2-binary
python-pptx